% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.service.R
\name{accountMetaData}
\alias{accountMetaData}
\title{Retrieve account details such as current tier, payment status, message quote usage, etc}
\usage{
accountMetaData()
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL}
}
\value{
parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Retrieve account details such as current tier, payment status, message quote usage, etc
}
