% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.service.R
\name{accountUsage}
\alias{accountUsage}
\title{Retrieve current month usage for your account.}
\usage{
accountUsage(type)
}
\arguments{
\item{type}{"messages" | "rules" | "rule-records" | alterts" | "alert-records"}

\item{endpoint}{a string which will form the variable are of the endpoint URL}
}
\value{
parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Retrieve current month usage for your account.
}
