% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{advancedStats}
\alias{advancedStats}
\title{Retrieve advancedStats detail for a symbol as dataframe}
\usage{
advancedStats(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 20 message units per call per symbol
}
\examples{
keyStats("AAPL")
}
