% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{balanceSheet}
\alias{balanceSheet}
\title{retrieve balance sheet for a symbol as dataframe}
\usage{
balanceSheet(symbol, period = "quarter", lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{period}{(quarter) "annual" | "quarter"}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve balance sheet for a symbol as dataframe
}
\examples{
balanceSheet("AAPL",period = "quarter", lastN =4)
}
