% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{batch_request}
\alias{batch_request}
\title{batch calls to market}
\usage{
batch_request(
  symbol = "market",
  types = NULL,
  symbols = NULL,
  range = NULL,
  ...
)
}
\arguments{
\item{symbol}{a market symbol or "market"}

\item{types}{a comma separated list of data types to return such as quote,chart,news}

\item{symbols}{if symbol is "market" then a commas separated list of market symbols}

\item{range}{if one of the types is chart, then the range parameter for the chart eq. 1d,1m,5y}

\item{...}{Parameters that are sent to individual endpoints can be specified in batch calls and will be applied to each supporting endpoint.}
}
\description{
batch calls to market
}
