% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{cashflowStatement}
\alias{cashflowStatement}
\title{retrieve cash flow statement for a symbol as dataframe}
\usage{
cashflowStatement(symbol, period = "quarter", lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{period}{(quarter) "annual" | "quarter"}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve cash flow statement for a symbol as dataframe
}
\examples{
cashflowStatement("AAPL",period = "quarter", lastN =4)
}
