% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{dailyTimeSeriesAdapter}
\alias{dailyTimeSeriesAdapter}
\title{Retrieve previous day adjusted price data for one or more stocks}
\usage{
dailyTimeSeriesAdapter(df)
}
\arguments{
\item{df}{dataframe returned from previousDay call}
}
\value{
a dataframe
}
\description{
Data Weighting: 2 message units per symbol
}
\examples{
previousDay("AAPL") \%>\% dailyTimeSeriesAdapter()
}
