% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_point.service.R
\name{data_point_keys}
\alias{data_point_keys}
\title{get all data-point keys for a symbol}
\usage{
data_point_keys(symbol)
}
\arguments{
\item{symbol}{a market symbol}
}
\description{
get all data-point keys for a symbol
}
\seealso{
Other data point service functions: 
\code{\link{data_point}()}
}
\concept{data point service functions}
