% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_point.service.R
\name{data_points}
\alias{data_points}
\title{data-point for symbol key}
\usage{
data_points(symbol, key)
}
\arguments{
\item{symbol}{a market symbol}

\item{key}{a data-point key}
}
\description{
data-point for symbol key
}
