% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.service.R
\name{deepAuction}
\alias{deepAuction}
\title{IEX Depth of Book}
\usage{
deepAuction(symbol)
}
\arguments{
\item{symbol}{a market symbol, one and only one symbol}
}
\description{
DEEP is used to receive real-time depth of book quotations direct from IEX.
The depth of book quotations received via DEEP provide an aggregated size
of resting displayed orders at a price and side, and do not indicate the
size or number of individual orders at any price level. Non-displayed orders
and non-displayed portions of reserve orders are not represented in DEEP.
}
\details{
DEEP also provides last trade price and size information.
Trades resulting from either displayed or non-displayed orders matching on IEX will be reported.
Routed executions will not be reported.

Data Weighting: FREE

Data Timing: REAL TIME

Data Schedule: Market hours 9:30am-4pm ET

Data Source(s): Consolidated Tape Investors Exchange
}
\seealso{
Other deep service functions: 
\code{\link{deepOfficialPrice}()},
\code{\link{deepTrades}()},
\code{\link{deep}()}
}
\concept{deep service functions}
