% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.service.R
\name{deepOfficialPrice}
\alias{deepOfficialPrice}
\title{The Official Price message is used to disseminate the IEX Official Opening and Closing Prices.}
\usage{
deepOfficialPrice(symbol)
}
\arguments{
\item{symbol}{a market symbol, one and only one symbol}
}
\description{
These messages will be provided only for IEX Listed Securities.
}
\details{
Data Weighting: FREE

Data Source(s): Investors Exchange
}
\examples{
deepOfficialPrice('AAPL')\\
}
