% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.service.R
\name{deepTrades}
\alias{deepTrades}
\title{Trade report messages are sent when an order on the IEX Order Book
is executed in whole or in part. DEEP sends a Trade report message
for every individual fill.}
\usage{
deepTrades(symbol)
}
\arguments{
\item{symbol}{a market symbol, one and only one symbol}
}
\description{
Data Weighting: FREE
}
\details{
Data Source(s): Investors Exchange
}
\examples{
deepTrades('AAPL')\\
}
\seealso{
Other deep service functions: \code{\link{deepAuction}},
  \code{\link{deepOfficialPrice}}, \code{\link{deep}}
}
\concept{deep service functions}
