% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{delayedQuote}
\alias{delayedQuote}
\title{Retrieve delayed quote for a symbol as dataframe}
\usage{
delayedQuote(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 1 per quote called
}
\details{
Data Timing: 15min delayed

Data Schedule: 4:30am - 8pm ET M-F when market is open
}
\examples{
delayedQuote("AAPL")
}
