% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.service.R
\name{disablePayAsYouGo}
\alias{disablePayAsYouGo}
\title{Used to disable Pay-as-you-go on your account.}
\usage{
disablePayAsYouGo()
}
\value{
parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Note when you turn off pay-as-you-go, there is a 30 second period before you can re-enable.
}
