% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{dividends}
\alias{dividends}
\title{Retrieve dividends detail for a symbol as dataframe}
\usage{
dividends(symbol, timePeriod = "3m")
}
\arguments{
\item{symbol}{stock symbol}

\item{timePeriod}{is one of "next" | "1m" | "3m" | "6m" | "ytd" | "1y" | "2y" | "5y"}
}
\value{
a dataframe
}
\description{
Data Weighting: 10 message unit per period returned
}
\details{
Data Schedule: Updated at 9am UTC every day
}
\examples{
dividends("AAPL")
}
