% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{earningsToday}
\alias{earningsToday}
\title{Retrieve Earnings data for a given company including the actual EPS, consensus, and fiscal period.
Earnings are available quarterly (last 4 quarters) and annually (last 4 years).}
\usage{
earningsToday(symbol, lastN = 1)
}
\value{
a dataframe
}
\description{
Data Weighting: 1000 mesaage units per symbol per period
}
\details{
Data Schedule: Updates at 9am, 11am, 12pm UTC every da
}
\examples{
earningsToday()
}
