% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{estimates}
\alias{estimates}
\title{Provides the latest consensus estimate for the next fiscal period}
\usage{
estimates(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
Data Weighting: 10000 message units per symbol per period
}
\details{
Data Schedule: Updates at 9am, 11am, 12pm UTC every day
}
\examples{
estimates("AAPL", lastN =4)
}
