% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{financials}
\alias{financials}
\title{retrieve financial statement for a symbol as dataframe}
\usage{
financials(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve financial statement for a symbol as dataframe
}
\examples{
financials("AAPL", lastN =4)
}
