% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{fundOwnership}
\alias{fundOwnership}
\title{Returns the top 10 fund holders,
meaning any firm not defined as buy-side or sell-side such as mutual funds,
pension funds, endowments, investment firms,
and other large entities that manage funds on behalf of others.}
\usage{
fundOwnership(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 10000 message units per symbol per period
}
\examples{
fundOwnership("AAPL")
}
