% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{getQuote}
\alias{getQuote}
\title{retrieve quote detail for a symbol as dataframe}
\usage{
getQuote(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve quote detail for a symbol as dataframe
}
\examples{
getQuote("AAPL")
}
