% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{historyFor}
\alias{historyFor}
\title{retrieve history for a stock over chosen time-period as dataframe}
\usage{
historyFor(symbol, timePeriod = "1m", chartCloseOnly = FALSE,
  chartIEXOnly = FALSE, chartLastN = 0, chartInterval = 1,
  changeFromClose = FALSE, chartReset = FALSE, chartSimplify = FALSE,
  date = "")
}
\arguments{
\item{symbol}{stock symbol}

\item{timePeriod}{is one of "dynamic" | "date" | "1d" | "1m" | "3m" | "6m" | "ytd" | "1y" | "2y" | "5y"}

\item{chartCloseOnly}{= FALSE,(1 mu/minute 50 max.) All ranges except 1d. Will return adjusted data only with keys date, close, and volume.}

\item{chartIEXOnly}{= FALSE,(free) Only for 1d. Limits the return of intraday prices to IEX only data.}

\item{chartLastN}{= 0, If passed, chart data will return the last N elements}

\item{chartInterval}{= 1,If passed, chart data will return every Nth element as defined by chartInterval}

\item{changeFromClose}{= FALSE,}

\item{chartReset}{= FALSE, If true, 1d chart will reset at midnight instead of the default behavior of 9:30am ET.}

\item{chartSimplify}{= FALSE, If true, runs a polyline simplification using the Douglas-Peucker algorithm. This is useful if plotting sparkline charts.}
}
\value{
a dataframe
}
\description{
retrieve history for a stock over chosen time-period as dataframe
}
\examples{
history("AAPL")
}
