% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexcloud.service.R
\name{iex_api}
\alias{iex_api}
\title{Perform a get request to an endpoint on the iexcloud server}
\usage{
iex_api(endpoint)
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL}
}
\value{
iex_api class parsed response data, this will usually be a list of key:value pairs from parsed json object
}
\description{
Perform a get request to an endpoint on the iexcloud server
}
