% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexcloudR.R
\docType{package}
\name{iexcloudR}
\alias{iexcloudR}
\alias{iexcloudR-package}
\title{iexcloudR: Wrapper for iexcloud Market Data Api}
\description{
Complete interface to the iexcloud market data api from 
    IEX Group Inc. The Investors Exchange.
}
\details{
icloudR: A wrapper for the IEX Cloud API

##What is IEX Cloud?

The IEXCloud is a micropayment Api for financial and stock quote data.
The Api is hosted by IEX Group Inc. \href{https://iextrading.com/}{The Investors Exchange}.
Using the Api requires \href{https://iexcloud.io/cloud-login#/register}{registration}.

There is a free tier for use during initial api exploration and application development.
\itemize{
\item \href{https://iexcloud.io/pricing/}{Pricing}
\item \href{https://iexcloud.io/docs/api/}{Api Documentation}
}

This is functional wrapper which returns data as tibbles, with a few exceptions.
See the package documentation for call signitures, parameter descriptions, and the
micropayment cost of each function call.Security Tokens

During resistration at icloud.io you were given security tokens required to access this API.
This wrapper will read those tokens from the .env file. The .env file must include the
following three key pairs: note(put .env in your .gitignore file)\preformatted{IEXCLOUD_VERSION = "stable"
IEXCLOUD_PRIVATE_KEY = "pk_.........."
IEXCLOUD_SECRET_KEY  = "sk_.........."
}
}
\author{
Bruce C. Schardt, \email{schardt.bruce.curtis@gmail.com}
}
\keyword{internal}
