% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexcloud.service.R
\name{iexRaw}
\alias{iexRaw}
\title{Perform a get request to an endpoint on the iexcloud server}
\usage{
iexRaw(endpoint)
}
\arguments{
\item{endpoint}{a string which will form the variable are of the endpoint URL
#' @return raw response object}
}
\description{
Perform a get request to an endpoint on the iexcloud server
}
