% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{insiderRoster}
\alias{insiderRoster}
\title{Returns Insider Roster}
\usage{
insiderRoster(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 5000 message units per symbol
}
\examples{
insiderRoster("AAPL")
}
