% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{insiderSummary}
\alias{insiderSummary}
\title{Returns Insider Summary}
\usage{
insiderSummary(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 5000 message units per symbol
}
\examples{
insiderSummary("AAPL")
}
