% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{insiderTransactions}
\alias{insiderTransactions}
\title{Returns Insider Transactions}
\usage{
insiderTransactions(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 50 message units per transaction
}
\examples{
insiderTransactions("AAPL")
}
