% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{largestTrades}
\alias{largestTrades}
\title{Retrieve 15 minute delayed, last sale eligible trades.}
\usage{
largestTrades(symbol)
}
\arguments{
\item{symbol}{a market symbol}
}
\value{
a dataframe
}
\description{
Data Weight: 1 message unit per trade returned
}
\examples{
largestTrades("mostactive")
}
