% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.service.R
\name{list_all_times_series}
\alias{list_all_times_series}
\title{return a data.frame with all the time series id and schema information}
\usage{
list_all_times_series()
}
\description{
return a data.frame with all the time series id and schema information
}
\seealso{
Other time-series service: 
\code{\link{time_series}()}
}
\concept{time-series service}
