% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{listOf}
\alias{listOf}
\title{Retrieve quotes for symbols on defined lists as dataframe}
\usage{
listOf(listType = "mostactive")
}
\arguments{
\item{listType}{one of   | "mostactive" | "gainers" | "losers" | "iexvolume" | "iexpercent" | "infocus";}
}
\value{
a dataframe
}
\description{
Retrieve quotes for symbols on defined lists as dataframe
}
\examples{
listOf("mostactive")
}
