% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{listSymbols}
\alias{listSymbols}
\title{retrieve listSymbols  as dataframe}
\usage{
listSymbols(symbolType)
}
\arguments{
\item{symbolType}{"symbols" | "iex" | "otc" | "mutual-fund" |}
}
\value{
a dataframe
}
\description{
retrieve listSymbols  as dataframe
}
\examples{
symbols("iex")
}
