% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{logoFor}
\alias{logoFor}
\title{Retrieve logo url for symbols on defined lists as string}
\usage{
logoFor(symbol)
}
\arguments{
\item{symbol}{a market symbol}
}
\value{
a string
}
\description{
Retrieve logo url for symbols on defined lists as string
}
\examples{
logo("AAPL")
}
