% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{marketOpen}
\alias{marketOpen}
\title{retrieve the official open and close for a given symbol.}
\usage{
marketOpen()
}
\value{
symbol a market symbol
}
\description{
retrieve the official open and close for a given symbol.
}
\examples{
marketOpen('AAPL')
}
