% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{marketOpenClose}
\alias{marketOpenClose}
\title{retrieve the official open and close for all IEX symbols in the market. 15 min delayed}
\usage{
marketOpenClose()
}
\value{
df
}
\description{
retrieve the official open and close for all IEX symbols in the market. 15 min delayed
}
\examples{
marketOpenClose()
}
