% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{marketVolume}
\alias{marketVolume}
\title{retrieve real time traded volume on U.S. markets.}
\usage{
marketVolume()
}
\value{
a dataframe
}
\description{
retrieve real time traded volume on U.S. markets.
}
\examples{
marketVolumn()
}
