% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{newsFor}
\alias{newsFor}
\title{retrieve real time traded volume on U.S. markets.}
\usage{
newsFor(subject, lastN = 10)
}
\value{
subject "market" | symbol
}
\description{
retrieve real time traded volume on U.S. markets.
}
