% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{ohlc}
\alias{ohlc}
\title{retrieve the official open and close for a given symbol.}
\usage{
ohlc(symbol)
}
\value{
symbol a market symbol
}
\description{
retrieve the official open and close for a given symbol.
}
\examples{
ohlc('AAPL')
}
