% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{peersOf}
\alias{peersOf}
\title{retrieves an array of peer symbols.}
\usage{
peersOf(symbol)
}
\value{
symbol a market symbol
}
\description{
retrieves an array of peer symbols.
}
\examples{
peers('AAPL')
}
