% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{previousDay}
\alias{previousDay}
\title{Retrieve previous day adjusted price data for one or more stocks}
\usage{
previousDay(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 2 message units per symbol
}
\examples{
previousDay("AAPL")
}
