% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{priceOf}
\alias{priceOf}
\title{retrieve current adjusted price data for one or more stocks}
\usage{
priceOf(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a number
}
\description{
retrieve current adjusted price data for one or more stocks
}
\examples{
priceOf("AAPL")
}
