% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{priveTarget}
\alias{priveTarget}
\title{retrieve the latest avg, high, and low analyst price target for a symbol.}
\usage{
priveTarget(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve the latest avg, high, and low analyst price target for a symbol.
}
\examples{
priceTarget("AAPL")
}
