% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{quoteFor}
\alias{quoteFor}
\title{retrieve quote detail for a symbol as dataframe}
\usage{
quoteFor(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
Data Weighting: 1 per quote called or streamed
}
\examples{
quoteFor("AAPL")
}
