% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{recommendationTrends}
\alias{recommendationTrends}
\title{Analyst Recommendation Summary}
\usage{
recommendationTrends(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
1000 message units per symbol
}
\examples{
recommendationTrends("AAPL")
}
