% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{socialSentiment}
\alias{socialSentiment}
\title{Retrieve Social Sentiment Data from StockTwits. Data can be viewed as a daily value, or by minute for a given date.}
\usage{
socialSentiment(symbol, date, type = "daily")
}
\arguments{
\item{symbol}{stock symbol}

\item{date}{as string YYYYMMDD, supposed to be optional, but is required at this time}

\item{type}{"daily" | "minute"}
}
\value{
a dataframe
}
\description{
Data Weighting:
\itemize{
\item 100 per symbol per date for daily sentiment
\item 200 per symbol per date for by minute sentiment
}
}
\details{
Data Timing: Realtime

Data Schedule: Continuous

Data Source(s): StockTwits
}
\examples{
socialSentiment("AAPL", "20190206","daily")
}
