% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{splits}
\alias{splits}
\title{Retrieve splits for a symbol}
\usage{
splits(symbol, date, type = "daily")
}
\arguments{
\item{symbol}{stock symbol}

\item{date}{as string YYYYMMDD, supposed to be optional, but is required at this time}

\item{period}{"next" | "1m" | "3m" | "6m" | "ytd" | "1y" | "2y" | "5y"}
}
\value{
a dataframe
}
\description{
Data Weighting: 10 per symbol per record
}
\details{
Data Schedule: Updated at 9am UTC every day

Data Source(s): EventVestor
}
\examples{
splits("AAPL", period = "1m")
}
