% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.service.R
\name{time_series}
\alias{time_series}
\title{return data for specified time series id}
\usage{
time_series(id, key = NULL, sub_key = NULL, from = NULL, to = NULL,
  on = NULL, last = NULL, first = NULL, filter = NULL)
}
\arguments{
\item{from}{Optional
Returns data on or after the given from date. Format YYYY-MM-DD. Used together with the to parameter to define a date range.}

\item{to}{Optional
Returns data on or before the given to date. Format YYYY-MM-DD}

\item{on}{Optional
Returns data on the given date. Format YYYY-MM-DD}

\item{last}{Optional
Returns the latest n number of records in the series}

\item{first}{Optional
Returns the first n number of records in the series}

\item{filter}{Optional
The standard filter parameter. Filters return data to the specified comma delimited list of keys (case-sensitive)}
}
\description{
return data for specified time series id
}
\seealso{
Other time-series service: \code{\link{list_all_times_series}}
}
\concept{time-series service}
