% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tops.service.R
\name{tops}
\alias{tops}
\title{IEX TOP of Book Quote Feed,}
\usage{
tops(symbol = "", securityType = NULL, sector = NULL)
}
\arguments{
\item{symbol}{a market symbol, when absent will return all market symbols}

\item{securityType}{| NULL filter}

\item{sector}{| NULL filter}
}
\value{
a data frame (filtered to remove symbols with no recorded sales i.e. lastSaleTime==0)
}
\description{
TOPS provides IEX’s aggregated best quoted bid and offer position in
near real time for all securities on IEX’s displayed limit order book.
TOPS is ideal for developers needing both quote and trade data.
}
\details{
Data Weighting: FREE

Data Timing: REAL TIME

Data Schedule: Market hours 9:30am-4pm ET

Data Source(s): Consolidated Tape Investors Exchange
}
\examples{
tops('AAPL')
}
