% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tops.service.R
\name{topsLast}
\alias{topsLast}
\title{IEX last sale price}
\usage{
topsLast(symbol = "")
}
\arguments{
\item{symbol}{a market symbol, when absent returns all market symbols}
}
\description{
Last provides trade data for executions on IEX.
It is a near real time, intraday API that provides IEX last sale price,
size and time. Last is ideal for developers that need a lightweight
stock quote.
Data Weighting: FREE
}
\details{
Data Timing: REAL TIME

Data Schedule: Market hours 9:30am-4pm ET

Data Source(s): Consolidated Tape Investors Exchange
}
\examples{
topsLast('AAPL')
}
