% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{upcomingEarnings}
\alias{upcomingEarnings}
\title{returns upcoming earnings reportDates for market}
\usage{
upcomingEarnings(symbol = "market")
}
\description{
returns upcoming earnings reportDates for market
}
