% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{volumeByVenue}
\alias{volumeByVenue}
\title{Returns 15 minute delayed and 30 day average consolidated volume percentage of a stock, by market.
This call will always return 13 values, and will be sorted in ascending order by current day trading volume percentage.}
\usage{
volumeByVenue(symbol)
}
\arguments{
\item{symbol}{a market symbol}
}
\description{
Data Weighting: 20 message units per call
}
\details{
Data Timing: 15 min delayed

Data Schedule: Updated during regular market hours 9:30am-4pm ET

Data Source(s): Consolidated Tape Investors Exchange
}
\examples{
volumeByVenue('AAPL')
}
