% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{balanceSheet}
\alias{balanceSheet}
\title{Pulls balance sheet data. Available quarterly (4 quarters) and annually (4 years)}
\usage{
balanceSheet(symbol, period = "quarter", lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{period}{(quarter) "annual" | "quarter"}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
Data Weighting: 3000 message units per symbol per period
}
\details{
Data Schedule: Updates at 8am, 9am UTC daily
}
\examples{
balanceSheet("AAPL",period = "quarter", lastN =4)
}
