% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{company}
\alias{company}
\title{retrieve company detail for a symbol as dataframe}
\usage{
company(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve company detail for a symbol as dataframe
}
\examples{
company("AAPL")
}
