% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_point.service.R
\name{data_point}
\alias{data_point}
\title{data-point for symbol key}
\usage{
data_point(symbol, key)
}
\arguments{
\item{symbol}{a market symbol}

\item{key}{a data-point key}
}
\description{
data-point for symbol key
}
\seealso{
Other data point service functions: \code{\link{data_point_keys}}
}
\concept{data point service functions}
