% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{dividends}
\alias{dividends}
\title{retrieve dividends detail for a symbol as dataframe}
\usage{
dividends(symbol, timePeriod = "3m")
}
\arguments{
\item{symbol}{stock symbol}

\item{timePeriod}{is one of "next" | "1m" | "3m" | "6m" | "ytd" | "1y" | "2y" | "5y"}
}
\value{
a dataframe
}
\description{
retrieve dividends detail for a symbol as dataframe
}
\examples{
dividends("AAPL")
}
