% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{earnings}
\alias{earnings}
\title{retrieve earnings statement for a symbol as dataframe}
\usage{
earnings(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve earnings statement for a symbol as dataframe
}
\examples{
earnings("AAPL", lastN =4)
}
