% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{earningsToday}
\alias{earningsToday}
\title{Returns earnings that will be reported today}
\usage{
earningsToday()
}
\value{
a dataframe
}
\description{
Data Weighting: 1000 mesaage units per symbol per period
}
\details{
Data Schedule: Updates at 9am, 11am, 12pm UTC every da
}
\examples{
earningsToday()
}
