% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{effectiveSpread}
\alias{effectiveSpread}
\title{retrieve effective spread for a symbol as dataframe}
\usage{
effectiveSpread(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve effective spread for a symbol as dataframe
}
\examples{
effectiveSpread("AAPL")
}
