% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{estimates}
\alias{estimates}
\title{retrieve estimates for a symbol as dataframe}
\usage{
estimates(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve estimates for a symbol as dataframe
}
\examples{
estimates("AAPL", lastN =4)
}
