% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{financials}
\alias{financials}
\title{Pulls income statement, balance sheet, and cash flow data from the most recent reported quarter.}
\usage{
financials(symbol, lastN = 1)
}
\arguments{
\item{symbol}{stock symbol}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
Data Weighting: 5000 message units per symbol per period
}
\examples{
financials("AAPL", lastN =4)
}
