% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexcloud.service.R
\name{iex_api}
\alias{iex_api}
\title{Perform a get request to an endpoint on the iexcloud server}
\usage{
iex_api(endpoint)
}
\arguments{
\item{endpoint}{a "url" object or a character string which will form the variable part of the endpoint URL
for a discussion of url objects see httr::build_url}
}
\value{
a list, class iex_api, with keys of "status,content,url,iexcloud_messages_used,response
}
\description{
Perform a get request to an endpoint on the iexcloud server
}
