% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{incomeStatement}
\alias{incomeStatement}
\title{retrieve income statement for a symbol as dataframe}
\usage{
incomeStatement(symbol, period = "quarter", lastN = 1)
}
\arguments{
\item{symbol}{stock}

\item{period}{(quarter) "annual" | "quarter"}

\item{lastN}{(1) number of periods to report}
}
\value{
a dataframe
}
\description{
retrieve income statement for a symbol as dataframe
}
\examples{
incomeStatement("AAPL",period = "quarter", lastN =4)
}
