% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{keyStats}
\alias{keyStats}
\title{retrieve keyStats detail for a symbol as dataframe}
\usage{
keyStats(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve keyStats detail for a symbol as dataframe
}
\examples{
keyStats("AAPL")
}
