% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{newsFor}
\alias{newsFor}
\title{Provides intraday news from over 3,000 global news sources including major publications,
regional media, and social.. This endpoint returns up to the last 50 articles}
\usage{
newsFor(subject, lastN = 5)
}
\arguments{
\item{subject}{the stock symbol}

\item{lastN}{the last N news items to return}
}
\value{
subject "market" | symbol
}
\description{
Data Weight: 1 message unit per call
}
\examples{
newsFor('AAPL')
}
