% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{newsFor}
\alias{newsFor}
\title{Retrieve real time traded volume on U.S. markets.}
\usage{
newsFor(subject, lastN = 5)
}
\value{
subject "market" | symbol
}
\description{
Data Weight: 1 message unit per call
}
\examples{
newsFor('AAPL')
}
