% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{peersOf}
\alias{peersOf}
\title{retrieves an array of peer symbols.}
\usage{
peersOf(symbol)
}
\value{
an array of market symbols
}
\description{
Data Weighting: 500 message units per call
}
\examples{
peers('AAPL')
}
