% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{previousDay}
\alias{previousDay}
\title{retrieve previous day adjusted price data for one or more stocks}
\usage{
previousDay(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a dataframe
}
\description{
retrieve previous day adjusted price data for one or more stocks
}
\examples{
previousDay("AAPL")
}
