% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.service.R
\name{priceOf}
\alias{priceOf}
\title{Retrieve A single number, being the IEX real time price, the 15 minute delayed market price, or the previous close price, is returned.}
\usage{
priceOf(symbol)
}
\arguments{
\item{symbol}{stock symbol}
}
\value{
a number
}
\description{
Data Weighting: 1 message unit per symbol
}
\examples{
priceOf("AAPL")
}
